refactor(ex-ray): cover the keepalive registration path and move read-back to tests
