+++ .github/workflows/ci.yaml
@@ -451,11 +451,8 @@ jobs:
       - name: Test ex-ray Go (ECH gate)
         run: cargo xtask run ex-ray-tests
 
-  # ex-ray's own Go tests on every OS it ships for. The `test-ex-ray` job above
-  # is linux-only because it also runs upstream v2ray-core suites; this one is
-  # scoped to `crates/ex-ray` so it can go wide cheaply. Without it the
-  # build-tagged per-OS socket code (keepalive_windows.go, keepalive_darwin.go)
-  # would never be compiled, let alone run, in CI.
+  # Runs the ex-ray-unit-tests target (build.yaml) on windows/darwin -- see there
+  # for why it is split from test-ex-ray.
   test-ex-ray-platforms:
     name: Test ex-ray (Go, ${{ matrix.os }})
     strategy:
