+++ crates/ex-ray/keepalive_windows.go
@@ -23,40 +23,3 @@ func setTCPKeepAlive(fd uintptr, p keepAliveParams) error {
 	}
 	return nil
 }
-
-func readTCPKeepAlive(fd uintptr) (keepAliveParams, bool, error) {
-	h := windows.Handle(fd)
-	on, err := windows.GetsockoptInt(h, windows.SOL_SOCKET, windows.SO_KEEPALIVE)
-	if err != nil {
-		return keepAliveParams{}, false, newError("failed to read SO_KEEPALIVE").Base(err)
-	}
-	idle, err := windows.GetsockoptInt(h, windows.IPPROTO_TCP, windows.TCP_KEEPIDLE)
-	if err != nil {
-		return keepAliveParams{}, false, newError("failed to read TCP_KEEPIDLE").Base(err)
-	}
-	intvl, err := windows.GetsockoptInt(h, windows.IPPROTO_TCP, windows.TCP_KEEPINTVL)
-	if err != nil {
-		return keepAliveParams{}, false, newError("failed to read TCP_KEEPINTVL").Base(err)
-	}
-	cnt, err := windows.GetsockoptInt(h, windows.IPPROTO_TCP, windows.TCP_KEEPCNT)
-	if err != nil {
-		return keepAliveParams{}, false, newError("failed to read TCP_KEEPCNT").Base(err)
-	}
-	idle32, err := sockoptInt32("TCP_KEEPIDLE", idle)
-	if err != nil {
-		return keepAliveParams{}, false, err
-	}
-	intvl32, err := sockoptInt32("TCP_KEEPINTVL", intvl)
-	if err != nil {
-		return keepAliveParams{}, false, err
-	}
-	cnt32, err := sockoptInt32("TCP_KEEPCNT", cnt)
-	if err != nil {
-		return keepAliveParams{}, false, err
-	}
-	return keepAliveParams{
-		IdleSeconds:     idle32,
-		IntervalSeconds: intvl32,
-		Probes:          cnt32,
-	}, on != 0, nil
-}
