+++ crates/ex-ray/keepalive_wiring_test.go
@@ -12,13 +12,16 @@ import (
 	"github.com/v2fly/v2ray-core/v5/transport/internet"
 )
 
-// Registered once for the whole binary: internet.RegisterDialerController
-// appends to a process-global slice with no removal, so registering inside a
-// test would leak into every later test and would double under -count=2.
-// Verified safe: no other test in this package dials and then asserts on socket
-// options.
+// Registered once for the whole binary through the real production entry point,
+// so the flag-to-live-controller path main() runs is covered rather than
+// re-implemented here. At this point flag defaults are assigned, so this
+// registers keepAliveDialerController({15,15,3}) -- exactly testParams().
+//
+// Once per binary is required: internet.RegisterDialerController appends to a
+// process-global slice with no removal, so registering inside a test would leak
+// into every later test and would double under -count=2.
 func TestMain(m *testing.M) {
-	common.Must(internet.RegisterDialerController(keepAliveDialerController(testParams())))
+	common.Must(registerTCPKeepAlive())
 	os.Exit(m.Run())
 }
 
@@ -79,12 +82,10 @@ func dialSystem(t *testing.T, dialer *internet.DefaultSystemDialer, sockopt *int
 	return tcpConn
 }
 
-// A dial through v2ray-core's own DialSystem must come out with every keepalive
-// option applied. This pins the two vendored behaviours the design leans on: a
-// registered controller still reaching the fd, and SocketConfig's keepalive
-// fields still producing net.Dialer.KeepAlive = -1 so Go's stdlib does not
-// overwrite them after connect. A subrepo bump breaking either must fail here
-// rather than silently restore the ~150s bound.
+// Pins the two vendored behaviours the README's "TCP keepalive" section
+// describes: a registered controller reaching the fd, and SocketConfig's fields
+// suppressing Go's post-connect defaults. A subrepo bump breaking either must
+// fail here.
 //
 // Probes is the load-bearing value: on linux and darwin the vendored
 // applyOutboundSocketOptions sets idle and interval from the SocketConfig
@@ -122,5 +123,39 @@ func TestDialSystemDisabledKeepAlive(t *testing.T) {
 	_, enabled := readParams(t, conn)
 	if enabled {
 		t.Error("SO_KEEPALIVE is on after a dial with the disable sentinel; tcp-keepalive=0 must leave keepalive fully off, including Go's default")
+	}
+}
+
+// registerTCPKeepAlive's client-mode body -- validate, build, register -- is
+// what main() runs at startup. TestMain covers that it runs at all; this pins
+// that the controller it installs actually reaches a dial.
+//
+// It swaps in a scratch system dialer so the registration is isolated, then
+// rebuilds the binary-wide state TestMain established. internet exposes no
+// getter for the effective dialer, so restoring means reconstructing it: a
+// fresh DefaultSystemDialer plus the same registerTCPKeepAlive call.
+func TestRegisterTCPKeepAliveInstallsController(t *testing.T) {
+	t.Cleanup(func() {
+		internet.UseAlternativeSystemDialer(nil)
+		common.Must(registerTCPKeepAlive())
+	})
+
+	internet.UseAlternativeSystemDialer(nil)
+	if err := registerTCPKeepAlive(); err != nil {
+		t.Fatalf("registerTCPKeepAlive: %v", err)
+	}
+
+	want := testParams()
+	conn := dialSystem(t, nil, &internet.SocketConfig{
+		TcpKeepAliveIdle:     want.IdleSeconds,
+		TcpKeepAliveInterval: want.IntervalSeconds,
+	})
+
+	got, enabled := readParams(t, conn)
+	if !enabled {
+		t.Error("SO_KEEPALIVE is off; registerTCPKeepAlive did not install a working controller")
+	}
+	if got.Probes != want.Probes {
+		t.Errorf("probe count = %d, want %d; the probe count is the one value only the registered controller can set", got.Probes, want.Probes)
 	}
 }
